Move promote_like_jnp to jax.test_util
